moved code to separate file
